feat: 完成CFD模式链路改造与任务单更新

+++ market-maker-core/src/main/java/com/exchange/marketmaker/service/impl/MmPerformanceServiceImpl.java
@@ -100,7 +100,7 @@ public class MmPerformanceServiceImpl implements com.exchange.marketmaker.servic
 
         // 评分结果
         performance.setScore(score);
-        performance.setIsQualified(isQualified);
+        performance.setIsQualified((int) isQualified);
 
         // 保存到数据库
         mmPerformanceMapper.insert(performance);
